Rename convs>conv and blocks>block
